Oppdatert dto format
