+++ libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/syntvedtakshistorikkservice/v1/dagpenger/BruktInntektsperiode.java
@@ -1,0 +1,15 @@
+package no.nav.testnav.libs.dto.syntvedtakshistorikkservice.v1.dagpenger;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+public class BruktInntektsperiode {
+    @JsonProperty
+    private String forsteManed;
+
+    @JsonProperty
+    private String sisteManed;
+}
