+++ libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/syntvedtakshistorikkservice/v1/dagpenger/Vedtaksperiode.java
@@ -1,0 +1,20 @@
+package no.nav.testnav.libs.dto.syntvedtakshistorikkservice.v1.dagpenger;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+
+@NoArgsConstructor
+@AllArgsConstructor
+public class Vedtaksperiode {
+    @JsonProperty
+    @JsonFormat(pattern = "yyyy-MM-dd")
+    private LocalDate fom;
+
+    @JsonProperty
+    @JsonFormat(pattern = "yyyy-MM-dd")
+    private LocalDate tom;
+}
