+++ libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/syntvedtakshistorikkservice/v1/dagpenger/Vilkaar.java
@@ -1,0 +1,15 @@
+package no.nav.testnav.libs.dto.syntvedtakshistorikkservice.v1.dagpenger;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+public class Vilkaar {
+    @JsonProperty
+    private String kode;
+
+    @JsonProperty
+    private String status;
+}
